requirements.txt
Flask-SQLAlchemy==3.0.2
boto3==1.27.0
pydantic==1.10.11
alembic==1.11.1
beautifulsoup4==4.12.2
hvac==1.1.1
apache-libcloud==3.8.0
fasteners==0.19
